Hide unsafe code to prevent the compiler from optimising it
The type definition for `alternatives` is a record, so a clever compiler (such as Flambda 2) can assume that values of this type can never be equal to `0`. This translates all `alts == noalts` checks into trivial `false` constants, eventually breaking down with a segfault when `noalts` is actually passed.
My proposed fix is the minimal one: hide the constant `0` from the compiler, so it can't optimise away.
The real fix is to use a variant definition. Example:
```ocaml
and alternative =
| Noalts
| Alts of {
  cutto_alts : alternative;
  program : prolog_prog;
  adepth : int;
  agoal_hd : constant;
  ogoal_arg : term;
  ogoal_args : term list;
  agid : UUID.t; [@trace]
  goals : goal list;
  stack : frame;
  trail : T.trail;
  state : State.t;
  clauses : clause list;
  next : alternative;
}
```
It's exactly the same runtime representation as the current one, but type-safe.